Mejora de ejemplos de Map Reduce con DataSet en vez de JavaRDD

+++ src/main/java/com/cursosdedesarrollo/ejemplos/entities/Person.java
@@ -3,10 +3,11 @@ package com.cursosdedesarrollo.ejemplos.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import scala.math.BigInt;
 
 @Data
 @AllArgsConstructor
 public class Person {
     private String name;
-    private Integer age;
+    private Long age;
 }
